requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
sentence-transformers
InstructorEmbedding
faiss-cpu
python-dotenv
